+++ library/src/test/java/com/evernote/android/job/DailyJobTest.java
@@ -1,5 +1,7 @@
 package com.evernote.android.job;
 
+import android.support.annotation.NonNull;
+
 import com.evernote.android.job.test.DummyJobs;
 import com.evernote.android.job.test.JobRobolectricTestRunner;
 import com.evernote.android.job.test.TestClock;
@@ -14,6 +16,7 @@ import org.junit.runners.MethodSorters;
 import java.util.Calendar;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 import static org.assertj.core.api.Java6Assertions.assertThat;
 
@@ -250,10 +253,48 @@ public class DailyJobTest extends BaseJobManagerTest {
         int normalJobId = DailyJob.schedule(DummyJobs.createBuilder(DummyJobs.SuccessJob.class), time, time);
         assertThat(manager().getAllJobRequests()).hasSize(2);
 
-        assertThat(nowJobId).isEqualTo(1);
         executeJob(nowJobId, Job.Result.SUCCESS);
 
         assertThat(manager().getAllJobRequests()).hasSize(1);
         assertThat(manager().getJobRequest(normalJobId)).isNotNull();
+    }
+
+    @Test
+    public void verifyRequirementsEnforcedSkipsJob() {
+        long time = 1L;
+
+        final AtomicBoolean atomicBoolean = new AtomicBoolean(true);
+
+        manager().addJobCreator(new JobCreator() {
+            @Override
+            public Job create(@NonNull String tag) {
+                return new DailyJob() {
+                    @NonNull
+                    @Override
+                    protected DailyJobResult onRunDailyJob(Params params) {
+                        atomicBoolean.set(false);
+                        return DailyJobResult.SUCCESS;
+                    }
+                };
+            }
+        });
+
+        int jobId = DailyJob.schedule(new JobRequest.Builder("any").setRequiresCharging(true).setRequirementsEnforced(true), time, time);
+        assertThat(manager().getAllJobRequests()).hasSize(1);
+
+        executeJob(jobId, Job.Result.SUCCESS);
+
+        assertThat(manager().getAllJobRequests()).hasSize(1);
+        assertThat(atomicBoolean.get()).isTrue();
+
+        // now verify that the job is called without the requirement
+        manager().cancelAll();
+        jobId = DailyJob.schedule(new JobRequest.Builder("any").setRequiresCharging(false).setRequirementsEnforced(true), time, time);
+        assertThat(manager().getAllJobRequests()).hasSize(1);
+
+        executeJob(jobId, Job.Result.SUCCESS);
+
+        assertThat(manager().getAllJobRequests()).hasSize(1);
+        assertThat(atomicBoolean.get()).isFalse();
     }
 }